[AC-2727] feat: widget modularization - scroll to bottom button (#310)
I divide the detailed tasks of modularization and work on them
separately.

## Changes

- Added ScrollToBottomButton component.
- Update structure of foundation components

ticket: [AC-2727]


[AC-2727]:
https://sendbird.atlassian.net/browse/AC-2727?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
