+++ src/foundation/components/DateSeparator/css.ts
@@ -1,0 +1,7 @@
+import { css } from '@linaria/core';
+
+export const dateSeparatorLabel = css`
+  margin: 0 16px;
+  display: flex;
+  white-space: nowrap;
+`;
