+++ src/foundation/components/DateSeparator/index.tsx
@@ -1,35 +1,20 @@
-import { cx, css } from '@linaria/atomic';
+import { cx } from '@linaria/atomic';
 import { styled } from '@linaria/react';
 import { format } from 'date-fns/format';
 import type { Locale } from 'date-fns/locale';
 
-import { Label } from './Label';
-import { bgColors, themedColorVars } from '../colors/css';
-import { useLocalProps } from '../hooks/useLocalProps';
-import { SBUFoundationProps } from '../types';
+import { dateSeparatorLabel } from './css';
+import { bgColors, themedColorVars } from '../../colors/css';
+import { useLocalProps } from '../../hooks/useLocalProps';
+import { SBUFoundationProps } from '../../types';
+import { Label } from '../Label';
 
-type Props = {
+type Props = SBUFoundationProps<{
   locale?: Locale;
   date?: Date;
   formatString?: string;
-
   separatorColor?: keyof typeof bgColors | string;
-};
-
-const Separator = styled.div<{ color?: string }>`
-  border: none;
-  height: 1px;
-  display: inline-block;
-  width: 100%;
-  ${({ color }) => (color ? `background-color: ${color};` : '')}
-`;
-
-const center = css`
-  margin: 0 16px;
-  display: flex;
-  white-space: nowrap;
-`;
-
+}>;
 export const DateSeparator = ({
   className,
   children,
@@ -37,15 +22,15 @@ export const DateSeparator = ({
   date = new Date(),
   formatString = 'MMMM dd, yyyy',
   separatorColor,
-  ...props
-}: SBUFoundationProps<Props>) => {
-  const localProps = useLocalProps(props);
+  testId = 'sendbird-date-separator',
+}: Props) => {
+  const localProps = useLocalProps({ testId });
   const colorClassName = separatorColor ? bgColors[separatorColor as keyof typeof bgColors] : bgColors.onbackground4;
 
   return (
     <div className={cx('sendbird-separator', themedColorVars, className)} {...localProps}>
       <Separator className={colorClassName} color={separatorColor && !colorClassName ? separatorColor : undefined} />
-      <div className={center}>
+      <div className={dateSeparatorLabel}>
         {children ?? (
           <Label type={'caption2'} color={'onbackground2'}>
             {format(date, formatString, { locale })}
@@ -56,3 +41,11 @@ export const DateSeparator = ({
     </div>
   );
 };
+
+const Separator = styled.div<{ color?: string }>`
+  border: none;
+  height: 1px;
+  display: inline-block;
+  width: 100%;
+  ${({ color }) => (color ? `background-color: ${color};` : '')}
+`;
