+++ src/foundation/components/Placeholder/Placeholder.loading.tsx
@@ -1,8 +1,8 @@
-import { PlaceHolderProps } from './types';
+import { PlaceholderProps } from './types';
 import { Icon } from '../Icon';
 import { Loader } from '../Loader';
 
-const PlaceholderLoading = ({ iconSize = 48, className }: PlaceHolderProps) => {
+const PlaceholderLoading = ({ iconSize = 48, className }: PlaceholderProps) => {
   return (
     <Loader size={iconSize} className={className}>
       <Icon type={'spinner'} size={iconSize} />
