+++ src/foundation/components/Placeholder/types.ts
@@ -1,6 +1,6 @@
 import type { ReactNode } from 'react';
 
-export type PlaceHolderProps = {
+export type PlaceholderProps = {
   className?: string;
   iconSize?: string | number;
   children?: ReactNode;
